requirements.txt
# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
torch>=2.0.0
gymnasium>=0.28.0

# Data acquisition
yfinance>=0.2.0

# Technical indicators (optional - we have built-in implementations)
# ta-lib>=0.4.0  # Optional: commented out, requires C compiler
# pandas-ta>=0.3.14b0  # Not needed - we compute indicators manually

# Metrics and evaluation
scipy>=1.10.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Experiment tracking
tensorboard>=2.13.0
pyyaml>=6.0

# Utilities
tqdm>=4.65.0
